pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-clipper"
version = "1.0.1"
description = "Capture text, screenshots, and OCR content to Obsidian via Local REST API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Meeran E Mandhini"}
]
keywords = ["obsidian", "clipboard", "screenshot", "ocr", "capture"]
classifiers = [
    "Development Status :: 1- Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Text Processing",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=25.0.0",
    "isort>=5.13.0",
    "mypy>=1.15.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
obsidian-clipper = "obsidian_clipper.clipper:main"

[project.urls]
Homepage = "https://github.com/emeeran/obsidian-clipper.git"
Documentation = "https://github.com/emeeran/obsidian-clipper.git#readme"
Repository = "https://github.com/emeeran/obsidian-clipper.git"
Issues = "https://github.com/emeeran/obsidian-clipper.git/issues"

[tool.hatch.build.targets.wheel]
packages = ["obsidian_clipper"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["obsidian_clipper"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=obsidian_clipper --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
source = ["obsidian_clipper"]
branch = true
omit = ["tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
